remove groupDetails2
